Clarify that some tx feature types are negations of each other.
GADTs can only provide positive evidence, and we sometimes need positive
evidence of things that are negations of each other.

For example: whether the era supports multi-asset or ada-only are
opposites of each other. We need to provide this positive evidence
in either direction when making a TxOutValue for example

data TxOutValue era where
     TxOutAdaOnly :: OnlyAdaSupportedInEra era -> ...
     TxOutValue   :: MultiAssetSupportedInEra era -> ...

So it is more helpful if our feature functions give us the evidence in
both directions for these cases, so specifically:

multiAssetSupportedInEra :: CardanoEra era
                         -> Either (OnlyAdaSupportedInEra era)
                                   (MultiAssetSupportedInEra era)

The same holds for tx fees: implicit vs explicit. Adjust the TxFee
representation so it follows the same pattern with positive and negative
evidence with a new type TxFeesImplicitInEra, and provide

txFeesExplicitInEra :: CardanoEra era
                    -> Either (TxFeesImplicitInEra era)
                              (TxFeesExplicitInEra era)